requirements.txt
flask
flask-sqlalchemy
psycopg2-binary # Driver for PostgreSQL
python-dotenv # Recommended for local development environment variables